Cleaning Up
I removed the folder that I made in an earlier test.

I also renamed the folder Enemies to Saving Duck Hurlington.  I did this
so that all source code could could into the src folder inside without
being confusing.  From now on, all source files should go in there.
This will make it easier for us to add files to our projects, since they
will all be in the same place.

If you had the enemies folder in your file path for an eclipse project,
you may have to change it to Saving Duck Hurlington now.  I am sorry,
but this will make it easier in the long term.