SDH 8.4 Smart Collisions
The rectangles I use to detect collisions with work correctly now. I
just need to put in logic to actually stop the sprite from walking onto
the rock, as well as if he spawns on a rock, it will move him(or just
have him spawn not on rocks :p).