Some Display Stuff
Got a pause menu working with appropriate inputs working as well.
We now have an image for the sword projectile, still need to get more
images for the other assets in the game.  It also appropriately rotates
the image based on the direction it's going (NOTE: make all projectile
images have square dimensions)